test: replace the env setups with fixtures
