+++ tests/integration/incremental_indexing/test_incremental_indexing.py
@@ -1,30 +1,15 @@
 import os
-import numpy as np
+
 from jina.flow import Flow
-from jina.proto import jina_pb2
-from jina.drivers.helper import array2pb
 from jina.executors import BaseExecutor
 from jina.executors.indexers.vector import NumpyIndexer
 from jina.executors.indexers.keyvalue import BinaryPbIndexer
+from tests.integration.incremental_indexing import random_workspace, get_duplicate_docs
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 
 
-def get_duplicate_docs(num_docs=10):
-    result = []
-    unique_set = set()
-    for idx in range(num_docs):
-        doc = jina_pb2.Document()
-        content = int(idx / 2)
-        doc.embedding.CopyFrom(array2pb(np.array([content])))
-        doc.text = f'I am doc{content}'
-        result.append(doc)
-        unique_set.add(content)
-    return result, len(unique_set)
-
-
-def test_incremental_indexing_vecindexers(tmpdir):
-    os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE'] = str(tmpdir)
+def test_incremental_indexing_vecindexers(random_workspace):
     total_docs = 10
     duplicate_docs, num_uniq_docs = get_duplicate_docs(num_docs=total_docs)
 
@@ -34,15 +19,12 @@ def test_incremental_indexing_vecindexers(tmpdir):
     with f:
         f.index(duplicate_docs)
 
-    with BaseExecutor.load(os.path.join(tmpdir, 'vec_idx.bin')) as vector_indexer:
+    with BaseExecutor.load(random_workspace / 'vec_idx.bin') as vector_indexer:
         assert isinstance(vector_indexer, NumpyIndexer)
         assert vector_indexer.size == num_uniq_docs
 
-    del os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE']
 
-
-def test_incremental_indexing_docindexers(tmpdir):
-    os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE'] = str(tmpdir)
+def test_incremental_indexing_docindexers(random_workspace):
     total_docs = 10
     duplicate_docs, num_uniq_docs = get_duplicate_docs(num_docs=total_docs)
 
@@ -52,40 +34,33 @@ def test_incremental_indexing_docindexers(tmpdir):
     with f:
         f.index(duplicate_docs)
 
-    with BaseExecutor.load(os.path.join(tmpdir, 'doc_idx.bin')) as doc_indexer:
+    with BaseExecutor.load(random_workspace / 'doc_idx.bin') as doc_indexer:
         assert isinstance(doc_indexer, BinaryPbIndexer)
         assert doc_indexer.size == num_uniq_docs
 
-    del os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE']
 
-
-def test_incremental_indexing_sequential_indexers(tmpdir):
-    os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE'] = str(tmpdir)
-
+def test_incremental_indexing_sequential_indexers(random_workspace):
     total_docs = 20
     duplicate_docs, num_uniq_docs = get_duplicate_docs(num_docs=total_docs)
 
     f = (Flow()
-         .add(uses=os.path.join(cur_dir, 'inc_vectorindexer.yml'), shards=1)
-         .add(uses=os.path.join(cur_dir, 'inc_docindexer.yml'), shards=1))
+         .add(uses=(cur_dir / 'inc_vectorindexer.yml'), shards=1)
+         .add(uses=(cur_dir / 'inc_docindexer.yml'), shards=1))
 
     with f:
         f.index(duplicate_docs[:10])
         f.index(duplicate_docs)
 
-    with BaseExecutor.load(os.path.join(tmpdir, 'vec_idx.bin')) as vector_indexer:
+    with BaseExecutor.load(random_workspace / 'vec_idx.bin') as vector_indexer:
         assert isinstance(vector_indexer, NumpyIndexer)
         assert vector_indexer._size == num_uniq_docs
 
-    with BaseExecutor.load(os.path.join(tmpdir, 'doc_idx.bin')) as doc_indexer:
+    with BaseExecutor.load(random_workspace / 'doc_idx.bin') as doc_indexer:
         assert isinstance(doc_indexer, BinaryPbIndexer)
         assert doc_indexer._size == num_uniq_docs
 
-    del os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE']
 
-
-def test_incremental_indexing_sequential_indexers_with_shards(tmpdir):
-    os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE'] = str(tmpdir)
+def test_incremental_indexing_sequential_indexers_with_shards(random_workspace):
     total_docs = 1000
     duplicate_docs, num_uniq_docs = get_duplicate_docs(num_docs=total_docs)
 
@@ -105,7 +80,7 @@ def test_incremental_indexing_sequential_indexers_with_shards(tmpdir):
 
     vect_idx_size = 0
     for shard_idx in range(num_shards):
-        save_abspath = os.path.join(tmpdir, f'vec_idx-{shard_idx + 1}', 'vec_idx.bin')
+        save_abspath = (random_workspace / f'vec_idx-{shard_idx + 1}' / 'vec_idx.bin')
         with BaseExecutor.load(save_abspath) as vector_indexer:
             assert isinstance(vector_indexer, NumpyIndexer)
             vect_idx_size += vector_indexer._size
@@ -113,17 +88,14 @@ def test_incremental_indexing_sequential_indexers_with_shards(tmpdir):
 
     doc_idx_size = 0
     for shard_idx in range(num_shards):
-        save_abspath = os.path.join(tmpdir, f'doc_idx-{shard_idx + 1}', 'doc_idx.bin')
+        save_abspath = (random_workspace / f'doc_idx-{shard_idx + 1}' / 'doc_idx.bin')
         with BaseExecutor.load(save_abspath) as doc_indexer:
             assert isinstance(doc_indexer, BinaryPbIndexer)
             doc_idx_size += doc_indexer._size
     assert doc_idx_size == num_uniq_docs
 
-    del os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE']
 
-
-def test_incremental_indexing_parallel_indexers(tmpdir):
-    os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE'] = str(tmpdir)
+def test_incremental_indexing_parallel_indexers(random_workspace):
     total_docs = 1000
     duplicate_docs, num_uniq_docs = get_duplicate_docs(num_docs=total_docs)
 
@@ -138,19 +110,16 @@ def test_incremental_indexing_parallel_indexers(tmpdir):
         f.index(duplicate_docs[:500])
         f.index(duplicate_docs)
 
-    with BaseExecutor.load(os.path.join(tmpdir, 'vec_idx.bin')) as vector_indexer:
+    with BaseExecutor.load((random_workspace / 'vec_idx.bin')) as vector_indexer:
         assert isinstance(vector_indexer, NumpyIndexer)
         assert vector_indexer._size == num_uniq_docs
 
-    with BaseExecutor.load(os.path.join(tmpdir, 'doc_idx.bin')) as doc_indexer:
+    with BaseExecutor.load((random_workspace / 'doc_idx.bin')) as doc_indexer:
         assert isinstance(doc_indexer, BinaryPbIndexer)
         assert doc_indexer._size == num_uniq_docs
 
-    del os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE']
 
-
-def test_incremental_indexing_parallel_indexers_with_shards(tmpdir):
-    os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE'] = str(tmpdir)
+def test_incremental_indexing_parallel_indexers_with_shards(random_workspace):
     total_docs = 1000
     duplicate_docs, num_uniq_docs = get_duplicate_docs(num_docs=total_docs)
 
@@ -177,7 +146,7 @@ def test_incremental_indexing_parallel_indexers_with_shards(tmpdir):
 
     vect_idx_size = 0
     for shard_idx in range(num_shards):
-        save_abspath = os.path.join(tmpdir, f'vec_idx-{shard_idx + 1}', 'vec_idx.bin')
+        save_abspath = (random_workspace / f'vec_idx-{shard_idx + 1}' / 'vec_idx.bin')
         with BaseExecutor.load(save_abspath) as vector_indexer:
             assert isinstance(vector_indexer, NumpyIndexer)
             vect_idx_size += vector_indexer._size
@@ -185,10 +154,8 @@ def test_incremental_indexing_parallel_indexers_with_shards(tmpdir):
 
     doc_idx_size = 0
     for shard_idx in range(num_shards):
-        save_abspath = os.path.join(tmpdir, f'doc_idx-{shard_idx + 1}', 'doc_idx.bin')
+        save_abspath = (random_workspace / f'doc_idx-{shard_idx + 1}' / 'doc_idx.bin')
         with BaseExecutor.load(save_abspath) as doc_indexer:
             assert isinstance(doc_indexer, BinaryPbIndexer)
             doc_idx_size += doc_indexer._size
     assert doc_idx_size == num_uniq_docs
-
-    del os.environ['JINA_TEST_INCREMENTAL_INDEX_WORKSPACE']
