+++ tests/integration/incremental_indexing/uniq_docindexer.yml
@@ -1,6 +1,8 @@
 !UniquePbIndexer
 components:
-  - !DocIDCache {}
+  - !DocIDCache
+    metas:
+      workspace: $JINA_TEST_INCREMENTAL_INDEX_WORKSPACE
   - !BinaryPbIndexer
     with:
       index_filename: doc.gz
